+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2022-10-06
+### Added
+- Organization member management
+### Fixed
+- Updated `Organization.create()` documentation to reflect Air API changes
+
 ## [2.2.0] - 2022-05-27
 ### Added
 - `AirApi.topologies.create()` now accepts an absolute path to a DOT file on the caller's local disk